Add files via upload
This bot uses  to monitor footage which can be accessed from your local directory of your pc , instantly identifying the presence of fire or smoke. But it doesn’t stop at detection. As soon as a hazard is identified, the bot immediately places an emergency mobile call, sending a FIRE ALERT to designated individuals or authorities.